pyproject.toml
[tool.poetry]
name = "check-out-ai"
version = "0.1.0"
description = ""
authors = ["inthree3 <42310616+inthree3@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
openai = "^1.30.5"
ipykernel = "^6.29.4"
python-dotenv = "^1.0.1"
requests = "^2.32.3"
datetime = "^5.5"
pandas = "^2.2.2"
openpyxl = "^3.1.3"
fastapi = {extras = ["all"], version = "^0.111.0"}
pillow = "^10.3.0"
boto3 = "^1.34.117"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
